refactor(sera-types): unify RuntimeError, additively extend variants (sera-8s91/runerr)
Delete sera-runtime::error::RuntimeError (local duplicate). Additively extend
sera-types::runtime::RuntimeError with Llm, Tool, Io, Json, Http variants.
Update sera-runtime::sera_errors From impl and tests to use the canonical type.
